Set medium rateto get brewery response

+++ src/main/java/com/ciandt/bootcamp/model/api/GetBreweryResponse.java
@@ -26,7 +26,7 @@ public class GetBreweryResponse {
 
     private String websiteUrl;
 
-    private Integer mediumRate;
+    private Double mediumRate;
 
 
 }
